requirements.txt
# Agent-Browser: merged with Jane (industrial agent, vision tools, demos)
playwright>=1.40.0
anthropic>=0.18.0
python-dotenv>=1.0.0
streamlit>=1.28.0
langchain-core>=0.2.0
langchain-anthropic>=0.1.0
langgraph>=0.0.20
